Converting values passed to --args to ArgXY variable
